Renamed Job() class to Cron() and renamed the source file for autoloader simplicity Moved PostWidget class definition into widgets directory for autoloader/namespace reasons.
